still have error transaction service

+++ src/main/java/com/se1889_jv/swp391/swpstart/repository/TransactionPaymentRepository.java
@@ -25,8 +25,8 @@ public interface TransactionPaymentRepository extends JpaRepository<TransactionP
     Page<TransactionPayment> filterTransactions(
             @Param("startDate") LocalDateTime startDate,
             @Param("endDate") LocalDateTime endDate,
-            @Param("minAmount") Integer minAmount,
-            @Param("maxAmount") Integer maxAmount,
+            @Param("minAmount") Double minAmount,
+            @Param("maxAmount") Double maxAmount,
             @Param("status") TransactionStatus status,
             Pageable pageable
